+++ settings.py
@@ -13,8 +13,8 @@ class Settings:
 
         # Bullet settings
         self.bullet_speed = 1.5
-        self.bullet_width = 3
-        self.bullet_height = 15
+        self.bullet_width = 35
+        self.bullet_height = 10
         self.bullet_color = (60, 60 ,60)
         self.bullets_allowed = 3